Rename the view user service to details
